v0.6.0: Refactor webview rendering into composable modules (refs #148)
