+++ src/extension.ts
@@ -102,6 +102,20 @@ import {
   normalizeIntentOverrideState,
   normalizeIntentOverrideText,
 } from './intentOverride';
+import { renderPanelClientScript } from './webview/panelClientScript';
+import {
+  renderChangesSinceCard,
+  renderDetailsCard,
+  renderEvidenceCard,
+  renderQuickActionsCard,
+  renderRecapCard,
+  renderRestorePackCard,
+  renderStatusCard,
+  renderTimelineCard,
+  renderTitledListCard,
+  renderTrustCenterCard,
+} from './webview/panelCards';
+import { PANEL_WEBVIEW_STYLE } from './webview/panelStyles';
 import {
   buildResumePathSteps,
   buildResumePathStorageKey,
@@ -4215,48 +4229,23 @@ function renderWebview(
       return `<section class="timeline-group"><h4>${escapeHtml(group.label)}</h4><ul>${items}</ul></section>`;
     })
     .join('');
-  const timelineCard = config.showTimeline
-    ? `<div class="card">
-      <h3>Timeline</h3>
-      <button type="button" class="secondary" data-action="toggleTimeline" aria-expanded="false" aria-controls="timeline-content">Show timeline</button>
-      <div id="timeline-content" hidden>
-        ${timelineGroupsHtml || '<p class="muted">No timeline entries captured yet.</p>'}
-      </div>
-    </div>`
-    : '';
+  const timelineCard = renderTimelineCard({
+    showTimeline: config.showTimeline,
+    timelineGroupsTrustedHtml: timelineGroupsHtml,
+  });
   const recapCard =
     recapDoneItems.length > 0 || recapPendingItems.length > 0 || Boolean(recapFirstAction)
-      ? `<div class="card recap-card">
-      <h3>Session Recap</h3>
-      <div class="recap-grid">
-        <section>
-          <h4>Done since last resume</h4>
-          <ul class="compact-list">${recapDoneList || '<li>None captured yet.</li>'}</ul>
-        </section>
-        <section>
-          <h4>Pending / blocked</h4>
-          <ul class="compact-list">${recapPendingList || '<li>No blocker captured.</li>'}</ul>
-        </section>
-        <section>
-          <h4>Next safe action</h4>
-          <p class="companion-primary">${escapeHtml(recapFirstAction || 'Refresh summary to regenerate first-action guidance.')}</p>
-          <div class="status-actions">
-            ${recapPrimaryNextActionButton}
-            <button type="button" class="secondary" data-action="copyNextSteps">Copy next steps</button>
-            <button type="button" class="secondary" data-action="sessionAddCheckpoint">Add note</button>
-            <button type="button" class="secondary" data-action="checkpointOpenList">List notes</button>
-          </div>
-        </section>
-      </div>
-    </div>`
+      ? renderRecapCard({
+          recapDoneListTrustedHtml: recapDoneList,
+          recapPendingListTrustedHtml: recapPendingList,
+          recapFirstAction: recapFirstAction ?? '',
+          recapPrimaryNextActionButtonTrustedHtml: recapPrimaryNextActionButton,
+        })
       : '';
   const changesSinceItems = (summary.changesSinceLastResume ?? [])
     .map((item) => `<li>${escapeHtml(item)}</li>`)
     .join('');
-  const changesSinceCard = `<div class="card">
-      <h3>Changes Since Last Time</h3>
-      <ul class="compact-list">${changesSinceItems || '<li>No changes captured.</li>'}</ul>
-    </div>`;
+  const changesSinceCard = renderChangesSinceCard(changesSinceItems);
   const nudgeCard =
     primaryNudge || secondaryNudge || nudgeSuppressionLabel
       ? `<div class="card">
@@ -4286,391 +4275,55 @@ function renderWebview(
     </div>`
       : '';
 
-  return `<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    ${cspMetaTag}
-    <style nonce="${nonce}">
-      :root {
-        --surface-bg: var(--vscode-editorWidget-background);
-        --surface-border: var(--vscode-panel-border);
-        --surface-muted: var(--vscode-descriptionForeground);
-        --surface-strong: var(--vscode-foreground);
-        --accent: var(--vscode-focusBorder);
-      }
-      body {
-        color: var(--surface-strong);
-        background: var(--vscode-editor-background);
-        font-family: var(--vscode-font-family);
-        font-size: var(--vscode-font-size);
-        line-height: 1.5;
-        padding: 16px;
-      }
-      .card {
-        border: 1px solid var(--surface-border);
-        border-radius: 12px;
-        padding: 14px;
-        margin-bottom: 14px;
-        background: var(--surface-bg);
-      }
-      ul {
-        padding-left: 20px;
-      }
-      h3 {
-        margin-top: 0;
-        margin-bottom: 10px;
-        display: flex;
-        align-items: center;
-        gap: 8px;
-      }
-      h3::before {
-        content: '';
-        width: 8px;
-        height: 8px;
-        border-radius: 999px;
-        background: var(--accent);
-      }
-      h4 {
-        margin-bottom: 8px;
-      }
-      a {
-        color: var(--vscode-textLink-foreground);
-      }
-      a:hover {
-        color: var(--vscode-textLink-activeForeground);
-      }
-      pre {
-        white-space: pre-wrap;
-        word-break: break-word;
-      }
-      .details-markdown {
-        line-height: 1.45;
-      }
-      .details-markdown > :first-child {
-        margin-top: 0;
-      }
-      .details-markdown > :last-child {
-        margin-bottom: 0;
-      }
-      .details-markdown p,
-      .details-markdown li {
-        overflow-wrap: anywhere;
-      }
-      .details-markdown code {
-        font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', monospace;
-      }
-      .kind {
-        color: var(--vscode-descriptionForeground);
-      }
-      .mode {
-        color: var(--vscode-descriptionForeground);
-        font-size: 13px;
-      }
-      .status-label {
-        font-weight: 700;
-        font-size: 13px;
-      }
-      .status-detail {
-        margin-top: 6px;
-      }
-      .status-actions {
-        margin-top: 10px;
-        display: flex;
-        flex-wrap: wrap;
-        gap: 8px;
-      }
-      .step-evidence {
-        margin-top: 6px;
-        display: flex;
-        gap: 6px;
-        flex-wrap: wrap;
-      }
-      .step-actions {
-        margin-top: 8px;
-      }
-      .step-action {
-        padding: 4px 10px;
-        font-size: 12px;
-      }
-      .step-advisory {
-        margin-top: 6px;
-        font-size: 12px;
-      }
-      .badge {
-        display: inline-block;
-        border: 1px solid var(--vscode-widget-border);
-        border-radius: 999px;
-        padding: 2px 8px;
-        font-size: 12px;
-        text-decoration: none;
-        color: inherit;
-      }
-      .badge.clickable {
-        cursor: pointer;
-      }
-      .badge.kind-url {
-        border-color: var(--vscode-textLink-foreground);
-      }
-      .badge.kind-file {
-        border-color: var(--vscode-charts-green);
-      }
-      .evidence-kind {
-        color: var(--vscode-descriptionForeground);
-      }
-      .evidence-target {
-        color: var(--vscode-descriptionForeground);
-      }
-      .extra-evidence {
-        display: none;
-      }
-      .evidence-list.show-more .extra-evidence {
-        display: list-item;
-      }
-      details summary {
-        cursor: pointer;
-      }
-      .show-more-btn {
-        margin-top: 8px;
-      }
-      .muted {
-        color: var(--surface-muted);
-      }
-      .resume-path-list {
-        margin-top: 10px;
-      }
-      .resume-path-item {
-        margin-bottom: 8px;
-      }
-      .resume-path-toggle {
-        display: flex;
-        align-items: center;
-        gap: 8px;
-        cursor: pointer;
-        font-weight: 600;
-      }
-      .resume-path-toggle input[type='checkbox'] {
-        width: 14px;
-        height: 14px;
-      }
-      .resume-path-detail {
-        margin: 4px 0 0 22px;
-        font-size: 12px;
-      }
-      .blocker-disabled-reason {
-        margin-top: 8px;
-      }
-      .restore-grid {
-        display: grid;
-        grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
-        gap: 8px;
-      }
-      button {
-        border: 1px solid var(--vscode-button-border, transparent);
-        background: var(--vscode-button-background);
-        color: var(--vscode-button-foreground);
-        border-radius: 8px;
-        padding: 8px 10px;
-      }
-      button:focus-visible {
-        outline: 2px solid var(--vscode-focusBorder);
-        outline-offset: 2px;
-      }
-      button.secondary {
-        background: transparent;
-        color: var(--vscode-editor-foreground);
-        border-color: var(--vscode-widget-border);
-      }
-      .restore-grid button {
-        text-align: left;
-        cursor: pointer;
-      }
-      .restore-grid button:disabled {
-        opacity: 0.5;
-        cursor: default;
-      }
-      .restore-note {
-        color: var(--vscode-descriptionForeground);
-        font-size: 12px;
-        margin-top: 8px;
-      }
-      .note-actions {
-        display: flex;
-        gap: 8px;
-        flex-wrap: wrap;
-      }
-      .note-actions button {
-        border-radius: 6px;
-        padding: 6px 10px;
-        cursor: pointer;
-      }
-      .timeline-group ul {
-        margin-top: 0;
-      }
-      .timeline-group li {
-        display: grid;
-        grid-template-columns: 70px 1fr;
-        gap: 8px;
-        margin-bottom: 6px;
-      }
-      .timeline-time {
-        color: var(--vscode-descriptionForeground);
-        font-size: 12px;
-        padding-top: 2px;
-      }
-      .timeline-row {
-        display: flex;
-        flex-direction: column;
-        gap: 2px;
-      }
-      .timeline-detail {
-        color: var(--vscode-descriptionForeground);
-        font-size: 12px;
-        overflow-wrap: anywhere;
-      }
-      .quick-actions {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 8px;
-      }
-      .quick-actions button {
-        min-width: 160px;
-      }
-      .action-group + .action-group {
-        margin-top: 10px;
-      }
-      .action-group h4,
-      .action-group h5 {
-        margin: 0 0 8px 0;
-        font-size: 12px;
-        text-transform: uppercase;
-        letter-spacing: 0.05em;
-        color: var(--surface-muted);
-      }
-      .companion-grid {
-        display: grid;
-        grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-        gap: 12px;
-      }
-      .companion-block {
-        border: 1px solid var(--vscode-widget-border);
-        border-radius: 10px;
-        padding: 12px;
-        background: var(--vscode-editor-background);
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-      }
-      .companion-block h4 {
-        margin-top: 0;
-        margin-bottom: 6px;
-        font-size: 12px;
-        text-transform: uppercase;
-        letter-spacing: 0.04em;
-        color: var(--surface-muted);
-      }
-      .companion-primary {
-        margin: 0 0 8px 0;
-        font-weight: 700;
-        line-height: 1.4;
-      }
-      .companion-kicker {
-        margin: 0;
-        font-size: 12px;
-        text-transform: uppercase;
-        letter-spacing: 0.05em;
-        color: var(--surface-muted);
-      }
-      .companion-meta {
-        margin: 0;
-        color: var(--surface-muted);
-      }
-      .intent-editor {
-        border: 1px solid var(--vscode-widget-border);
-        border-radius: 8px;
-        padding: 8px;
-      }
-      .intent-editor-row {
-        margin-top: 6px;
-      }
-      .intent-editor input[type='text'] {
-        width: 100%;
-        box-sizing: border-box;
-        padding: 6px 8px;
-        border-radius: 6px;
-        border: 1px solid var(--vscode-input-border, var(--vscode-widget-border));
-        background: var(--vscode-input-background);
-        color: var(--vscode-input-foreground);
-      }
-      .intent-editor-actions {
-        display: flex;
-        gap: 8px;
-        margin-top: 8px;
-        flex-wrap: wrap;
-      }
-      .compact-list {
-        margin: 0 0 10px 0;
-        padding-left: 18px;
-      }
-      .companion-restore-grid {
-        display: grid;
-        grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
-        gap: 6px;
-      }
-      .companion-restore-grid button {
-        text-align: left;
-      }
-      .trust-row {
-        margin-bottom: 8px;
-      }
-      .trust-key {
-        font-weight: 600;
-      }
-      .recap-card .recap-grid {
-        display: grid;
-        grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-        gap: 10px;
-      }
-      .recap-card h4 {
-        margin-top: 0;
-      }
-    </style>
-  </head>
-  <body>
-    <div class="card">
-      <h3>Status</h3>
-      <div class="status-label">${escapeHtml(sourceLabel)} · ${escapeHtml(generatedAtLabel)}</div>
-      <div class="status-detail muted">${escapeHtml(statusHint)}</div>
-      <div class="status-detail"><strong>${escapeHtml(autoSummaryStatusLabel)}</strong></div>
-      <div class="status-detail muted">${escapeHtml(autoSummaryStatusDetail)}</div>
-      <div class="status-actions">
-        <button type="button" data-action="refreshSummary">Refresh summary now</button>
-        <button type="button" class="secondary" data-action="toggleAutoSummaries" ${autoSummaryToggleDisabledAttr}>${escapeHtml(autoSummaryToggleLabel)}</button>
-      </div>
-    </div>
-
-    <div class="card">
-      <h3>Trust Center</h3>
-      <div class="trust-row"><span class="trust-key">Tracking:</span> ${escapeHtml(trustTrackingLabel)}</div>
-      <div class="trust-row"><span class="trust-key">Stored locally:</span> ${escapeHtml(storedLocallyLabel)}</div>
-      <div class="trust-row"><span class="trust-key">Sent to AI:</span> ${escapeHtml(sentToAiLabel)}</div>
-      <div class="trust-row"><span class="trust-key">Based on:</span> ${escapeHtml(trustCue.headline.replace('Based on: ', ''))}</div>
-      <details>
-        <summary><strong>Why am I seeing this?</strong></summary>
-        <ul class="compact-list">${trustCueDetails || '<li>No evidence counts yet.</li>'}</ul>
-      </details>
-      <div class="status-actions">
-        <button type="button" class="secondary" data-action="toggleAutoSummaries" ${autoSummaryToggleDisabledAttr}>${escapeHtml(autoSummaryToggleLabel)}</button>
-        <button type="button" class="secondary" data-action="openPrivacySafety">Open Privacy & Safety</button>
-      </div>
-    </div>
-
-    ${checkpointCard}
-    ${scratchpadCard}
-    ${recapCard}
-    ${changesSinceCard}
-    ${nudgeCard}
-    ${renderResumeStackCard({
+  const statusCard = renderStatusCard({
+    sourceLabel,
+    generatedAtLabel,
+    statusHint,
+    autoSummaryStatusLabel,
+    autoSummaryStatusDetail,
+    autoSummaryToggleDisabledAttr,
+    autoSummaryToggleLabel,
+  });
+  const trustCenterCard = renderTrustCenterCard({
+    trustTrackingLabel,
+    storedLocallyLabel,
+    sentToAiLabel,
+    trustBasedOn: trustCue.headline.replace('Based on: ', ''),
+    trustCueDetailsTrustedHtml: trustCueDetails,
+    autoSummaryToggleDisabledAttr,
+    autoSummaryToggleLabel,
+  });
+  const nextStepsCard = renderTitledListCard({
+    title: 'Next Steps',
+    listItemsTrustedHtml: nextSteps,
+    emptyMessage: 'No next steps captured yet.',
+  });
+  const topFilesCard = renderTitledListCard({
+    title: 'Top Files',
+    listItemsTrustedHtml: topFiles,
+    emptyMessage: 'None captured',
+  });
+  const topLinksCard = renderTitledListCard({
+    title: 'Top Links / Files',
+    listItemsTrustedHtml: linkItems,
+    emptyMessage: 'None captured',
+  });
+  const quickActionsCard = renderQuickActionsCard(quickActionGroups);
+  const restorePackCard = renderRestorePackCard(restorePackGroups, trusted);
+  const evidenceCard = renderEvidenceCard({
+    evidenceItemsTrustedHtml: evidenceItems,
+    hasExtraEvidence,
+  });
+  const detailsCard = renderDetailsCard(detailsHtml);
+  const bodyCards = [
+    statusCard,
+    trustCenterCard,
+    checkpointCard,
+    scratchpadCard,
+    recapCard,
+    changesSinceCard,
+    nudgeCard,
+    renderResumeStackCard({
       intent: summary.intent,
       intentOverridden: summary.intentOverridden,
       intentEditorTrustedHtml: intentEditorHtml,
@@ -4691,321 +4344,34 @@ function renderWebview(
       blockerDisabledReasonTrustedHtml: blockerDisabledReasonHtml,
       blockerActionTrustedHtml: blockerActionHtml,
       restoreSectionsTrustedHtml: companionRestoreSections,
-    })}
-    ${resumePathCard}
+    }),
+    resumePathCard,
+    confidenceCard,
+    nextStepsCard,
+    topFilesCard,
+    topLinksCard,
+    timelineCard,
+    quickActionsCard,
+    restorePackCard,
+    evidenceCard,
+    detailsCard,
+  ]
+    .filter(Boolean)
+    .join('\n\n');
+  const clientScript = renderPanelClientScript(MAX_INTENT_OVERRIDE_CHARS);
 
-    ${confidenceCard}
-
-    <div class="card">
-      <h3>Next Steps</h3>
-      <ul>${nextSteps}</ul>
-    </div>
-
-    <div class="card">
-      <h3>Top Files</h3>
-      <ul>${topFiles || '<li>None captured</li>'}</ul>
-    </div>
-
-    <div class="card">
-      <h3>Top Links / Files</h3>
-      <ul>${linkItems || '<li>None captured</li>'}</ul>
-    </div>
-
-    ${timelineCard}
-
-    <div class="card">
-      <h3>Quick Actions</h3>
-      ${quickActionGroups}
-    </div>
-
-    <div class="card">
-      <h3>Restore Pack</h3>
-      ${restorePackGroups}
-      ${
-        trusted
-          ? ''
-          : '<div class="restore-note">Restricted Mode: task/debug/branch execution actions are disabled.</div>'
-      }
-    </div>
-
-    <div class="card">
-      <details>
-        <summary><strong>Evidence</strong></summary>
-        <ul class="evidence-list" id="evidence-list">${evidenceItems || '<li>None captured</li>'}</ul>
-        ${
-          hasExtraEvidence
-            ? '<button type="button" class="show-more-btn" data-action="toggleEvidenceMore">Show more</button>'
-            : ''
-        }
-      </details>
-    </div>
-
-    <div class="card">
-      <h3>Details</h3>
-      <div class="details-markdown">${detailsHtml}</div>
-    </div>
+  return `<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    ${cspMetaTag}
+    <style nonce="${nonce}">${PANEL_WEBVIEW_STYLE}</style>
+  </head>
+  <body>
+    ${bodyCards}
 
     <script nonce="${nonce}">
-      const vscode = acquireVsCodeApi();
-      const hostActions = new Set([
-        'fixSummary',
-        'checkpointPinToggle',
-        'checkpointMarkDone',
-        'checkpointDismiss',
-        'checkpointOpenList',
-        'openScratchpad',
-        'appendScratchpad',
-        'setScratchpadScope',
-        'sessionAddCheckpoint',
-        'setIntentOverride',
-        'clearIntentOverride',
-        'copyNextSteps',
-        'copySummary',
-        'copyPromptAndOpenCodex',
-        'refreshSummary',
-        'toggleAutoSummaries',
-        'acknowledgeNudge',
-        'dismissNudge',
-        'openPrivacySafety',
-        'rateHelpfulness',
-        'runNextStepAction',
-        'restoreWorkingSet',
-        'restoreJumpToLastEdit',
-        'restoreReopenFiles',
-        'restoreOpenChangedFiles',
-        'restoreRerunTask',
-        'restoreRerunDebug',
-        'restoreOpenProblems',
-        'restoreOpenDiagnosticFile',
-        'restoreCheckoutPreviousBranch',
-        'restoreCopyFailingCommand'
-      ]);
-      const viewState = Object.assign(
-        { evidenceExpanded: false, timelineExpanded: false },
-        vscode.getState() || {},
-      );
-
-      function persistViewState() {
-        vscode.setState(viewState);
-      }
-
-      function setEvidenceExpanded(expanded) {
-        const list = document.getElementById('evidence-list');
-        const toggle = document.querySelector('[data-action="toggleEvidenceMore"]');
-        if (!(list instanceof HTMLElement) || !(toggle instanceof HTMLElement)) {
-          viewState.evidenceExpanded = false;
-          persistViewState();
-          return;
-        }
-
-        list.classList.toggle('show-more', expanded);
-        toggle.textContent = expanded ? 'Show less' : 'Show more';
-        viewState.evidenceExpanded = expanded;
-        persistViewState();
-      }
-
-      function setTimelineExpanded(expanded) {
-        const timeline = document.getElementById('timeline-content');
-        const toggle = document.querySelector('[data-action="toggleTimeline"]');
-        if (!(timeline instanceof HTMLElement) || !(toggle instanceof HTMLElement)) {
-          viewState.timelineExpanded = false;
-          persistViewState();
-          return;
-        }
-
-        if (expanded) {
-          timeline.removeAttribute('hidden');
-          toggle.textContent = 'Hide timeline';
-          toggle.setAttribute('aria-expanded', 'true');
-        } else {
-          timeline.setAttribute('hidden', 'true');
-          toggle.textContent = 'Show timeline';
-          toggle.setAttribute('aria-expanded', 'false');
-        }
-        viewState.timelineExpanded = expanded;
-        persistViewState();
-      }
-
-      setEvidenceExpanded(Boolean(viewState.evidenceExpanded));
-      setTimelineExpanded(Boolean(viewState.timelineExpanded));
-
-      function parseDatasetInteger(rawValue) {
-        if (typeof rawValue !== 'string') {
-          return undefined;
-        }
-        const parsed = Number(rawValue);
-        if (!Number.isInteger(parsed) || parsed < 0) {
-          return undefined;
-        }
-        return parsed;
-      }
-
-      const maxIntentOverrideChars = ${MAX_INTENT_OVERRIDE_CHARS};
-      function normalizeIntentOverrideInput(rawValue) {
-        if (typeof rawValue !== 'string') {
-          return undefined;
-        }
-
-        const normalized = rawValue
-          .replace(/\\r?\\n/g, ' ')
-          .replace(/\\s+/g, ' ')
-          .trim()
-          .slice(0, maxIntentOverrideChars);
-        return normalized || undefined;
-      }
-
-      document.addEventListener('change', (event) => {
-        const target = event.target;
-        if (!(target instanceof HTMLElement)) {
-          return;
-        }
-
-        const toggle = target.closest('[data-resume-path-toggle="true"]');
-        if (!(toggle instanceof HTMLInputElement)) {
-          return;
-        }
-
-        const stepId = toggle.dataset.resumePathStepId;
-        if (
-          stepId !== 'confirmIntent' &&
-          stepId !== 'runNextSafeAction' &&
-          stepId !== 'clearBlocker'
-        ) {
-          return;
-        }
-
-        vscode.postMessage({
-          type: 'resumePathToggle',
-          stepId,
-          completed: toggle.checked,
-        });
-      });
-
-      document.addEventListener('keydown', (event) => {
-        if (!(event.target instanceof HTMLInputElement)) {
-          return;
-        }
-
-        if (event.target.id !== 'intent-override-input') {
-          return;
-        }
-
-        if (event.key !== 'Enter') {
-          return;
-        }
-
-        event.preventDefault();
-        const intent = normalizeIntentOverrideInput(event.target.value);
-        if (!intent) {
-          return;
-        }
-
-        vscode.postMessage({
-          type: 'setIntentOverride',
-          intent,
-        });
-      });
-
-      document.addEventListener('click', (event) => {
-        const target = event.target;
-        if (!(target instanceof HTMLElement)) {
-          return;
-        }
-
-        const actionElement = target.closest('[data-action]');
-        if (actionElement instanceof HTMLElement) {
-          event.preventDefault();
-          const action = actionElement.dataset.action;
-          if (typeof action !== 'string' || !action) {
-            vscode.postMessage({ type: 'blockedLink' });
-            return;
-          }
-
-          if (action === 'toggleEvidenceMore') {
-            setEvidenceExpanded(!Boolean(viewState.evidenceExpanded));
-            return;
-          }
-
-          if (action === 'toggleTimeline') {
-            setTimelineExpanded(!Boolean(viewState.timelineExpanded));
-            return;
-          }
-
-          if (action === 'openEvidence') {
-            const evidenceId = actionElement.dataset.evidenceId?.trim();
-            if (!evidenceId) {
-              vscode.postMessage({ type: 'blockedLink' });
-              return;
-            }
-            vscode.postMessage({ type: 'openEvidence', evidenceId });
-            return;
-          }
-
-          if (action === 'openLink') {
-            const index = parseDatasetInteger(actionElement.dataset.linkIndex);
-            if (index === undefined) {
-              vscode.postMessage({ type: 'blockedLink' });
-              return;
-            }
-            vscode.postMessage({ type: 'openLink', index });
-            return;
-          }
-
-          if (action === 'openTopFile') {
-            const index = parseDatasetInteger(actionElement.dataset.topFileIndex);
-            if (index === undefined) {
-              vscode.postMessage({ type: 'blockedLink' });
-              return;
-            }
-            vscode.postMessage({ type: 'openTopFile', index });
-            return;
-          }
-
-          if (action === 'runNextStepAction') {
-            const stepIndex = parseDatasetInteger(actionElement.dataset.stepIndex);
-            if (stepIndex === undefined) {
-              vscode.postMessage({ type: 'blockedLink' });
-              return;
-            }
-            const primarySurface = actionElement.dataset.primaryNextSafeAction;
-            if (primarySurface === 'home' || primarySurface === 'recap') {
-              vscode.postMessage({ type: 'runNextStepAction', stepIndex, primarySurface });
-              return;
-            }
-            vscode.postMessage({ type: 'runNextStepAction', stepIndex });
-            return;
-          }
-
-          if (action === 'setIntentOverride') {
-            const input = document.getElementById('intent-override-input');
-            if (!(input instanceof HTMLInputElement)) {
-              return;
-            }
-
-            const intent = normalizeIntentOverrideInput(input.value);
-            if (!intent) {
-              input.focus();
-              return;
-            }
-            vscode.postMessage({ type: 'setIntentOverride', intent });
-            return;
-          }
-
-          if (hostActions.has(action)) {
-            vscode.postMessage({ type: action });
-            return;
-          }
-
-          vscode.postMessage({ type: 'blockedLink' });
-          return;
-        }
-
-        const anchor = target.closest('a');
-        if (anchor) {
-          event.preventDefault();
-          vscode.postMessage({ type: 'blockedLink' });
-        }
-      });
+${clientScript}
     </script>
   </body>
 </html>`;
